requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
httpx==0.28.1
PyMuPDF==1.25.1
pytesseract==0.3.13
Pillow==11.1.0
